chore: Move modules in middleware and add mirrors for backward compatibility
